requirements.txt
python-multipart
sqlalchemy
pyjwt
fastapi
python-dotenv
uvicorn
mypy
google-cloud-storage
bcrypt
psycopg2-binary
pydantic